+++ src/backend/s3/wado.rs
@@ -1,6 +1,6 @@
 use crate::api::wado::{
-	InstanceResponse, RenderedRequest, RenderedResponse, RenderingRequest, RetrieveError,
-	RetrieveInstanceRequest, WadoService,
+	InstanceResponse, MetadataRequest, RenderedRequest, RenderedResponse, RenderingRequest,
+	RetrieveError, RetrieveInstanceRequest, WadoService,
 };
 use crate::backend::dimse::cmove::movescu::MoveError;
 use crate::config::{S3Config, S3EndpointStyle};
@@ -120,6 +120,10 @@ impl WadoService for S3WadoService {
 	}
 
 	async fn render(&self, _request: RenderingRequest) -> Result<RenderedResponse, RetrieveError> {
+		unimplemented!()
+	}
+
+	async fn metadata(&self, request: MetadataRequest) -> Result<InstanceResponse, RetrieveError> {
 		unimplemented!()
 	}
 }